eliminando FireBaseConfiguration e importando y exportando links por medio de firebaseMain.js, renombrando main.js a routes.js , actualizando c ontroller.js con las promesas directas de firebaseMain.js y renombrando componentes
