+++ src/index.html
@@ -6,14 +6,14 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <link rel="stylesheet" type="text/css" href="./styles.css">
-    <title>Document</title>
+    <title>Help Potatoes</title>
 </head>
 
 <body>
 
     <div id="root" class="rootClass">
     </div>
-    <script type="module" src="main.js"></script>
+    <script type="module" src="routes.js"></script>
 </body>
 
 </html>
