+++ src/lib/controller.js
@@ -1,12 +1,51 @@
-import { registerWithEmailFb, updateProfilefb, sendEmailfb, registerWithGoogleFb, loginFb, logOutfb, onGetPost, savePostfb } from "./firebaseMain.js";
-import { onNavigate } from "../main.js";
+import {
+    getAuth,
+    app,
+    db,
+    createUserWithEmailAndPassword,
+    updateProfile,
+    sendEmailVerification,
+    GoogleAuthProvider,
+    signInWithPopup,
+    signInWithEmailAndPassword,
+    signOut,
+
+    //firestore
+
+    collection,
+    addDoc,
+    serverTimestamp,
+    getDocs,
+    onSnapshot,
+    query,
+    orderBy
+
+} from "./firebaseMain.js";
+
+import { onNavigate } from "../routes.js";
 import { loginEvents, homeEvents } from "../DOMevents.js";
 
+const auth = getAuth(app);
+
 export const registerWithEmail = (email, password, name, photo) => {
-    registerWithEmailFb(email, password)
+    return createUserWithEmailAndPassword(auth, email, password)
+        .then((userCredential) => {
+            // Signed in
+            const user = userCredential.user;
+            console.log(user);
+        })
         .then(() => {
-            updateProfilefb(name, photo);
-            sendEmailfb()
+            updateProfile(auth.currentUser, {
+                displayName: name,
+                photoURL: photo,
+            }).then(() => {
+                // Profile updated!
+                console.log('ya actualice tu nombre y foto');
+            }).catch((error) => {
+                // An error occurred
+                console.log('error al actualizar');
+            });
+            sendEmailVerification(auth.currentUser)
                 .then(() => {
                     const showModalEmailVerification = document.getElementById('modalEmailV');
                     const hideModal = document.getElementById('closeModal');
@@ -33,29 +72,89 @@ export const registerWithEmail = (email, password, name, photo) => {
 
 
 export const registerWithGoogle = (photo) => {
-    registerWithGoogleFb(photo)
-        .then((res) => {
+    const provider = new GoogleAuthProvider();
+    return signInWithPopup(auth, provider)
+        .then((result) => {
+            updateProfile(auth.currentUser, {
+                photoURL: photo,
+            }).then(() => {
+                // Profile updated!
+                console.log('ya actualice tu nombre y foto de google');
+            }).catch((error) => {
+                // An error occurred
+                console.log('error al actualizar');
+            });;
+            // This gives you a Google Access Token. You can use it to access the Google API.
+            const credential = GoogleAuthProvider.credentialFromResult(result);
+            const token = credential.accessToken;
+            // The signed-in user info.
+            const user = result.user;
+            console.log(user);
+            const userData = {
+                userName: user.displayName,
+                userState: user.emailVerified,
+                userPhoto: user.photoURL,
+                userUid: user.uid
+            }
+            return userData;
+
+        }).then(() => {
             onNavigate('/home');
             const userProfile = document.getElementById('profileContainer');
             const newPost = document.getElementById('postContainer');
 
             userProfile.innerHTML = `<div class="containerPhotoProfile">
-            <img src="./img/pape.png" alt="" class="iconUserDefault">
-            </div>
-            `
+                <img src="./img/pape.png" alt="" class="iconUserDefault">
+                </div>
+                `
             newPost.innerHTML = `
-            <input type="text" class="postInput" placeholder="¡Hola! ¿Qué quieres compartir?">
-            <button class="btnShare">Publicar</button>`;
+                <input type="text" class="postInput" placeholder="¡Hola! ¿Qué quieres compartir?">
+                <button class="btnShare">Publicar</button>`;
             seePost();
+            // console.log('llego hasta aquí')
             homeEvents();
-        })
+        }).catch((error) => {
+            // Handle Errors here.
+            const errorCode = error.code;
+            const errorMessage = error.message;
+            // The email of the user's account used.
+            const email = error.email;
+            // The AuthCredential type that was used.
+            const credential = GoogleAuthProvider.credentialFromError(error);
+        });
 
 };
+
 //función de cerrar sesión
-export const logOut = () => logOutfb();
+export const logOut = () => {
+    console.log('estoy entrando');
+    return signOut(auth)
+        .then(() => {
+            console.log('Se esta cerrando la sesion')
+                // Sign-out successful.
+        }).catch((error) => {
+            const errorCode = error.code;
+            const errorMessage = error.message;
+            console.log(errorMessage);
+            // An error happened.
+        });
+}
 
 export const login = (email, password) => {
-    loginFb(email, password)
+    return signInWithEmailAndPassword(auth, email, password)
+        .then((userCredential) => {
+            // Signed in
+            const user = userCredential.user;
+            console.log(user);
+            console.log('logueado con exito')
+            const userData = {
+                userName: user.displayName,
+                userState: user.emailVerified,
+                userPhoto: user.photoURL,
+                userUid: user.uid
+            }
+            return userData;
+        })
         .then((res) => {
             console.log(res)
             if (res.userState === false) {
@@ -98,10 +197,30 @@ export const login = (email, password) => {
 };
 
 //función de guardar post en firestore 
-export const savePost = (input) => savePostfb(input);
+export const savePost = (input) => {
+    const user = auth.currentUser;
+    console.log(user);
+    return addDoc(collection(db, "post"), {
+        id: user.uid,
+        name: user.displayName,
+        email: user.email,
+        post: input,
+        photoURL: user.photoURL,
+        date: serverTimestamp()
+    }).then((docRef) => {
+        console.log("Tu post ya esta en la nube", docRef.id);
+    }).catch((e) => {
+        console.error("Error adding document: ", e);
+
+    });
+}
+
+//AÑADIENDO AQUÍ !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+//funciónpara actualización de post en timepo real
+export const onGetPost = (callback) => onSnapshot(query(collection(db, "post"), orderBy('date', 'desc')), callback);
 
 export const seePost = () => {
-    onGetPost((query) => {
+    return onGetPost((query) => {
         document.getElementById('postPublic').innerHTML = '';
         query.forEach((doc) => {
             document.getElementById('postPublic').innerHTML += `
